Modified things to no avail.
The class entities in the Demo don't get passed through the
transformer at all for some strange unknown reason.
This means the fetch methods don't work.  I don't understand
as it worked before.  And works for the Tranformer.main programme.

Am going to try and revert back to when it did and see if I can get
it going from there.